requirements.txt
altair
anyio
appnope
argon2-cffi
argon2-cffi-bindings
arrow==1.3.0
asn1crypto==1.5.1
asttokens
attrs
beautifulsoup4==4.12.3
bleach
blinker
cachetools
certifi
cffi
charset-normalizer
click
cloudpickle
comm
cryptography
debugpy
decorator
defusedxml
exceptiongroup
executing
filelock
gitdb
GitPython
idna
ipykernel
ipython
jedi
Jinja2
jsonpointer
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab_code_formatter
jupyterlab_pygments
markdown-it-py
MarkupSafe
matplotlib-inline
mdurl
mistune
nbclient
nbconvert
nbformat
nest-asyncio
numpy
overrides
packaging
pandas
pandocfilters
parso
pexpect
pillow
platformdirs
prometheus_client
prompt_toolkit
protobuf
psutil
ptyprocess
pyarrow
pycparser
pydeck
Pygments
PyJWT
pyOpenSSL
python-dateutil
python-json-logger
pytz
PyYAML
pyzmq
referencing
requests
rfc3339-validator
rfc3986-validator
rich
rpds-py
Send2Trash
six
smmap
sniffio
snowflake-connector-python
snowflake-snowpark-python
sortedcontainers
soupsieve
sqlparse
streamlit==1.39.0
tenacity
terminado
tinycss2
toml
tomlkit
toolz
tornado
traitlets
typing_extensions
tzdata
urllib3
wcwidth
webencodings
websocket-client
